(PC-33723)[API] chore: move AdageCulturalPartner(s) to core/educational/schemas.py
