+++ api/src/pcapi/core/educational/api/adage.py
@@ -13,6 +13,8 @@ from pcapi.core.educational import models as educational_models
 from pcapi.core.educational import repository as educational_repository
 from pcapi.core.educational.api import address as address_api
 from pcapi.core.educational.api.venue import get_relative_venues_by_siret
+from pcapi.core.educational.schemas import AdageCulturalPartner
+from pcapi.core.educational.schemas import AdageCulturalPartners
 from pcapi.core.mails.transactional import send_eac_offerer_activation_email
 from pcapi.core.offerers import models as offerers_models
 from pcapi.core.offerers import repository as offerers_repository
@@ -37,9 +39,7 @@ def find_collective_bookings_for_adage(
     )
 
 
-def get_cultural_partners(
-    *, since_date: datetime | None = None, force_update: bool = False
-) -> venues_serialize.AdageCulturalPartners:
+def get_cultural_partners(*, since_date: datetime | None = None, force_update: bool = False) -> AdageCulturalPartners:
     CULTURAL_PARTNERS_CACHE_KEY = "api:adage_cultural_partner:cache"
     CULTURAL_PARTNERS_CACHE_TIMEOUT = 24 * 60 * 60  # 24h in seconds
 
@@ -61,7 +61,7 @@ def get_cultural_partners(
 
     cultural_partners_json = typing.cast(str, cultural_partners_json)
     cultural_partners = json.loads(cultural_partners_json)
-    return parse_obj_as(venues_serialize.AdageCulturalPartners, {"partners": cultural_partners})
+    return parse_obj_as(AdageCulturalPartners, {"partners": cultural_partners})
 
 
 def get_cultural_partner(siret: str) -> venues_serialize.AdageCulturalPartnerResponseModel:
@@ -102,7 +102,7 @@ def get_venue_by_id_for_adage_iframe(
     return venue, relative
 
 
-def synchronize_adage_ids_on_offerers(partners_from_adage: list[venues_serialize.AdageCulturalPartner]) -> None:
+def synchronize_adage_ids_on_offerers(partners_from_adage: list[AdageCulturalPartner]) -> None:
     adage_sirens: set[str] = {p.siret[:9] for p in partners_from_adage if (p.actif == 1 and p.siret)}
     existing_sirens: dict[str, bool] = dict(
         offerers_models.Offerer.query.filter(offerers_models.Offerer.siren != None).with_entities(
