+++ api/src/pcapi/core/educational/schemas.py
@@ -1,4 +1,40 @@
+from datetime import datetime
+
 from pcapi.routes.serialization import BaseModel
+
+
+class AdageCulturalPartner(BaseModel):
+    id: int
+    venueId: int | None
+    siret: str | None
+    regionId: int | None
+    academieId: str | None
+    statutId: int | None
+    labelId: int | None
+    typeId: int | None
+    communeId: str | None
+    libelle: str
+    adresse: str | None
+    siteWeb: str | None
+    latitude: float | None
+    longitude: float | None
+    statutLibelle: str | None
+    labelLibelle: str | None
+    typeIcone: str | None
+    typeLibelle: str | None
+    communeLibelle: str | None
+    communeDepartement: str | None
+    academieLibelle: str | None
+    regionLibelle: str | None
+    domaines: str | None
+    actif: int | None
+    dateModification: datetime
+    synchroPass: int | None
+    domaineIds: str | None
+
+
+class AdageCulturalPartners(BaseModel):
+    partners: list[AdageCulturalPartner]
 
 
 class RedactorInformation(BaseModel):
